futuredsp: Move TapsAccessor to own file
