+++ futuredsp/src/fir.rs
@@ -2,10 +2,8 @@ use core::cmp::Ordering;
 #[cfg(not(RUSTC_IS_STABLE))]
 use core::intrinsics::{fadd_fast, fmul_fast};
 
+use crate::TapsAccessor;
 use num_complex::Complex;
-
-extern crate alloc;
-use alloc::vec::Vec;
 
 /// Represents the status of a computation.
 #[derive(PartialEq, Eq, Clone, Copy, Debug)]
@@ -50,57 +48,6 @@ pub trait FirKernel<SampleType>: Send {
         input: &[SampleType],
         output: &mut [SampleType],
     ) -> (usize, usize, ComputationStatus);
-}
-
-pub trait TapsAccessor: Send {
-    type TapType;
-
-    fn num_taps(&self) -> usize;
-
-    /// Gets the `index`th tap.
-    ///
-    /// # Safety
-    /// The invariant `index < num_taps()` must be upheld.
-    unsafe fn get(&self, index: usize) -> Self::TapType;
-}
-
-impl<const N: usize> TapsAccessor for [f32; N] {
-    type TapType = f32;
-
-    fn num_taps(&self) -> usize {
-        N
-    }
-
-    unsafe fn get(&self, index: usize) -> f32 {
-        debug_assert!(index < self.num_taps());
-        *self.get_unchecked(index)
-    }
-}
-
-impl<const N: usize> TapsAccessor for &[f32; N] {
-    type TapType = f32;
-
-    fn num_taps(&self) -> usize {
-        N
-    }
-
-    unsafe fn get(&self, index: usize) -> f32 {
-        debug_assert!(index < self.num_taps());
-        *self.get_unchecked(index)
-    }
-}
-
-impl TapsAccessor for Vec<f32> {
-    type TapType = f32;
-
-    fn num_taps(&self) -> usize {
-        self.len()
-    }
-
-    unsafe fn get(&self, index: usize) -> f32 {
-        debug_assert!(index < self.num_taps());
-        *self.get_unchecked(index)
-    }
 }
 
 /// A non-resampling FIR filter. Calling `work()` on this struct always
