+++ futuredsp/src/lib.rs
@@ -2,3 +2,6 @@
 #![cfg_attr(not(RUSTC_IS_STABLE), feature(core_intrinsics))]
 
 pub mod fir;
+
+mod tapsaccessor;
+pub use tapsaccessor::TapsAccessor;
